Initial Upload
For starters we have a script that is used to pull data from an API
based on its station ID. This is performed in the dataPull.php file and
stored in the database.

Secondly we have the dbconnect.php file which is used by the rest of the
files to connect to the database in a simple file.

The other thing is findStation.php which is currently being worked on.
It pulls the loc/lat from the user and can then be used to get the users
nearest station.