Updated index.php
Big update, the index.php file has had its looks altered and now it will
follow this style throughout the website.

We have also tied the findStation.php script to the search bar so that
the user can have their location found and the station automatically
entered.

Currently working on allowing the user to search through the stations.