Searching Updated
Searching now reveals panel's containing google maps image of the
location in question as well as the name of the station. This can then
be linked to the final page that shows the required information.

The search boxes needs moving down to make sure they dont cover the
search bar as well as making sure they dont go over the next section.